swtichdev: disable multicast snooping from bridge by default
Some of our regressions [e.g., spider69] exhibit intermittent failures
we've used to consider are related to time it takes the setup to load.
Such examples are l2-004-bridge_bond, l2-010-bridge_vlan1d_sanity.

Examining the setup in question, it looks like the first couple of IPv6
packets are failing to pass. But contrary to our initial assumption, it
doesn't look related to link - the physical carrier is seen is up tens
of seconds before traffic commences.

Examining the traffic, we can see the sender is sending ICMP6, neighbor
solicitation packets for a couple of seconds prior to the receiver
getting them. Our revised assumption is that it might be that due to
ordering issues when creating the setup topology the switch's bridge
is constructed too late and missse MLD sent by the receiver; As a
result, the later solicitation packets would not be directed to the
receiver.

Since all of these tests aren't really meant to test multicast behavior
for our device, instead of trying to fight with the topology simply
disable the multicast snooping on the switch's bridge - that would
provide flood behavior which would suffice for our tests.

Signed-off-by: Yuval Mintz <yuvalm@mellanox.com>