recipes: switchdev: ipip-006: Test decap routes at down tunnels
When a tunnel is down, a local route should never be offloaded as a
decap of that tunnel. Test that it is actually so.

In the same area, also test that downing a tunnel removes the decap
offload, and that upping a tunnel with a decap route installs the decap
offload.

Signed-off-by: Petr Machata <petrm@mellanox.com>