recipes: switchdev: ipip: Ping from proper iface
To verify that slow-path traps work, ipip-004 test should send GRE
packets with the wrong GRE key to a switch termination address. The
theory is that since the HW can't forward those on the fast path, we
should see these packets propagated to slow path. However, due to using
the wrong interface to ping from, the ping packets are never actually
wrapped, and the traffic that gets trapped are just ARP requests. That
shows as RX counter test passing, while the TX fails, because there are
no ICMP destination unreachable responses.

The situation is similar for ipip-006, where decap-only flow is tested
by sending encapsulated packets from the second host, and the first host
responds through another channel. Again, to actually send encapsulated
packets, a proper interface needs to be selected.

Fix this by performing the ping on the GRE interface itself.

Signed-off-by: Petr Machata <petrm@mellanox.com>