recipes: switchdev: Test RED-ECN on two children
Create a bottleneck for traffic by setting the ingress port to be in higher
rate the then egress port.
Set PRIO with 2 bands (3 are configured but only 2 are used).
Set RED on one band and RED-ECN on the other.
Test the first band in the same way that the basic RED test does.
Test the second band in the same way that the basic RED-ECN test does.
Send the traffic to the wanted band by changing it's TOS field.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>