recipes: switchdev: Test PRIO remapping with RED
(Like test qos-013 but with remapping instead of changing TOS)

Create a bottleneck for traffic by setting the ingress port to be in higher
rate the then egress port.
Set PRIO with 2 bands (3 are configured but only 2 are used).
Set on each band RED with different configuration.
Test each band in the same way that the basic RED test does.
Send the traffic to the wanted band by changing PRIO's mapping of the
traffic to a different band.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>