recipes: switchdev: Test PRIO remapping
(Like test qos-012 but with remapping instead of changing TOS)

Create a bottleneck for traffic by setting the ingress port to be in higher
rate the then egress port.
Set PRIO of 3 bands on the egress port.
Send traffic in the middle band in a rate that will consume all the
bandwidth.
While this traffic run:
* Send traffic in the lowest priority band and see that it doesn't being
  sent out.
* Remap this traffic to the highest priority band and see that it is
  received.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>