recipes: switchdev: Test removing RED as child by parent remapping
Set PRIO of 8 bands on a port.
Set RED on band 7. Check that it is set.
Change PRIO to have fewer bands.
See that the traffic that was previously with RED is now without it.
Change PRIO back to its previous config and see that RED hadn't come back.
Set RED on band 7 and see that its stats starts with 0's.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>